backend/requirements.txt
flask==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
flask-cors==4.0.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

